including popovers in closing functionality when clicking away
I had to change the trigger for the popover to manual to get better
control over when and how the popover is shown. Some issues were
happening where you had to double click on the preferences button to
get it to show. It seemed to be caused by the popover just getting out
of sync. Manually hiding and closing seems to fix the problem.