functionality for closing the playlist menu without having to use the close button (#206)
* adding track list menu

* small fix for previous and next funtionality

* removing the use of var and updating the track list accordingly when a track is selected

* adding bootstrap tooltips to icons

* adding tooltip for play button

* Update title from track list to playlist

* adding functionality for closing the track list without having to use the close button

* including popovers in closing functionality when clicking away

I had to change the trigger for the popover to manual to get better
control over when and how the popover is shown. Some issues were
happening where you had to double click on the preferences button to
get it to show. It seemed to be caused by the popover just getting out
of sync. Manually hiding and closing seems to fix the problem.